+++ parser/gjson_extend.go
@@ -179,7 +179,7 @@ func (c *GjsonExtendMetric) GetDate(key string) (t time.Time) {
 }
 
 func (c *GjsonExtendMetric) GetDateTime(key string) (t time.Time) {
-	if v := c.GetInt(key); v != 0 {
+	if v := c.GetFloat(key); v != 0 {
 		return time.Unix(int64(v), int64(v*1e9)%1e9)
 	}
 
@@ -189,7 +189,7 @@ func (c *GjsonExtendMetric) GetDateTime(key string) (t time.Time) {
 }
 
 func (c *GjsonExtendMetric) GetDateTime64(key string) (t time.Time) {
-	if v := c.GetInt(key); v != 0 {
+	if v := c.GetFloat(key); v != 0 {
 		return time.Unix(int64(v), int64(v*1e9)%1e9)
 	}
